add: package refactor and fixes navigation
