+++ app/src/main/java/dev/baseio/discordjetpackcompose/ui/routes/dashboard/DashboardRoute.kt
@@ -1,0 +1,23 @@
+package dev.baseio.discordjetpackcompose.ui.routes.dashboard
+
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.composable
+import androidx.navigation.navigation
+import dev.baseio.discordjetpackcompose.navigator.ComposeNavigator
+import dev.baseio.discordjetpackcompose.navigator.DiscordRoute
+import dev.baseio.discordjetpackcompose.navigator.DiscordScreen
+import dev.baseio.discordjetpackcompose.ui.routes.dashboard.main.DashboardScreen
+
+fun NavGraphBuilder.dashboardRoute(
+    composeNavigator: ComposeNavigator,
+) {
+    navigation(
+        startDestination = DiscordScreen.Dashboard.name,
+        route = DiscordRoute.Dashboard.name
+    ) {
+        composable(DiscordScreen.Dashboard.name) {
+            DashboardScreen(composeNavigator = composeNavigator)
+        }
+    }
+
+}
