+++ app/src/main/java/dev/baseio/discordjetpackcompose/ui/routes/dashboard/components/SearchBox.kt
@@ -1,4 +1,4 @@
-package dev.baseio.discordjetpackcompose.ui.components
+package dev.baseio.discordjetpackcompose.ui.routes.dashboard.components
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.background
