+++ app/src/main/java/dev/baseio/discordjetpackcompose/ui/routes/dashboard/components/ServerIconSelector.kt
@@ -1,4 +1,4 @@
-package dev.baseio.discordjetpackcompose.ui.components
+package dev.baseio.discordjetpackcompose.ui.routes.dashboard.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxHeight
