+++ app/src/main/java/dev/baseio/discordjetpackcompose/ui/routes/dashboard/components/ServerIconSelectorItem.kt
@@ -1,4 +1,4 @@
-package dev.baseio.discordjetpackcompose.ui.components
+package dev.baseio.discordjetpackcompose.ui.routes.dashboard.components
 
 import androidx.annotation.DrawableRes
 import androidx.compose.animation.core.animateDpAsState
