+++ app/src/main/java/dev/baseio/discordjetpackcompose/ui/routes/onboarding/screens/register/RegisterScreen.kt
@@ -1,6 +1,5 @@
 package dev.baseio.discordjetpackcompose.ui.routes.onboarding.screens.register
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -37,7 +36,7 @@ import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import dev.baseio.discordjetpackcompose.entities.CountryEntity
 import dev.baseio.discordjetpackcompose.navigator.ComposeNavigator
-import dev.baseio.discordjetpackcompose.ui.components.CountryPicker
+import dev.baseio.discordjetpackcompose.navigator.DiscordRoute
 import dev.baseio.discordjetpackcompose.ui.components.DiscordScaffold
 import dev.baseio.discordjetpackcompose.ui.theme.DiscordJetpackComposeTheme
 import dev.baseio.discordjetpackcompose.ui.utils.Strings
@@ -47,7 +46,7 @@ import kotlinx.coroutines.launch
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
 fun RegisterScreen(
-    composeNavigator: ComposeNavigator? = null,
+    composeNavigator: ComposeNavigator,
     registrationViewModel: RegistrationViewModel = hiltViewModel()
 ) {
 
@@ -127,7 +126,9 @@ fun RegisterScreen(
                         contentColor = Color(0xFFFFFFFF),
                     ),
                     shape = RoundedCornerShape(4.dp),
-                    onClick = { },
+                    onClick = {
+                        composeNavigator.navigate(DiscordRoute.Dashboard.name)
+                    },
                 ) {
                     Text(stringResource(id = Strings.next))
                 }
@@ -151,20 +152,6 @@ fun RegisterScreen(
                 countrySearchQuery = countrySearchQuery,
                 onQueryUpdated = { updatedQuery -> countrySearchQuery = updatedQuery }
             )
-        }
-    }
-}
-
-@Preview
-@Composable
-fun PreviewRegisterScreen() {
-    DiscordJetpackComposeTheme {
-        // A surface container using the 'background' color from the theme
-        Surface(
-            modifier = Modifier.fillMaxSize(),
-            color = DiscordColorProvider.colors.background
-        ) {
-            RegisterScreen()
         }
     }
 }
